Creating 3D Models by Yourself
I have tried many methods to solve the display issue, but none of them have worked. I suspect the problem might be with the 3D models generated by EasyEDA or KiCad. Therefore, I attempted to create a 3D model using images of the front and back of the PCB, but it still didn't resolve the issue.